+++ src/RcppExports.cpp
@@ -54,16 +54,16 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// hermite_integral_val_quantile_adap
-NumericMatrix hermite_integral_val_quantile_adap(int N, NumericVector x, NumericMatrix hermite_function_mat);
-RcppExport SEXP _hermiter_hermite_integral_val_quantile_adap(SEXP NSEXP, SEXP xSEXP, SEXP hermite_function_matSEXP) {
+// hermite_integral_val_upper
+NumericMatrix hermite_integral_val_upper(int N, NumericVector x, NumericMatrix hermite_function_mat);
+RcppExport SEXP _hermiter_hermite_integral_val_upper(SEXP NSEXP, SEXP xSEXP, SEXP hermite_function_matSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< int >::type N(NSEXP);
     Rcpp::traits::input_parameter< NumericVector >::type x(xSEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type hermite_function_mat(hermite_function_matSEXP);
-    rcpp_result_gen = Rcpp::wrap(hermite_integral_val_quantile_adap(N, x, hermite_function_mat));
+    rcpp_result_gen = Rcpp::wrap(hermite_integral_val_upper(N, x, hermite_function_mat));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -113,7 +113,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_hermiter_hermite_normalization", (DL_FUNC) &_hermiter_hermite_normalization, 1},
     {"_hermiter_hermite_function", (DL_FUNC) &_hermiter_hermite_function, 3},
     {"_hermiter_hermite_integral_val", (DL_FUNC) &_hermiter_hermite_integral_val, 3},
-    {"_hermiter_hermite_integral_val_quantile_adap", (DL_FUNC) &_hermiter_hermite_integral_val_quantile_adap, 3},
+    {"_hermiter_hermite_integral_val_upper", (DL_FUNC) &_hermiter_hermite_integral_val_upper, 3},
     {"_hermiter_hermite_int_full_domain", (DL_FUNC) &_hermiter_hermite_int_full_domain, 1},
     {"_hermiter_standardizeInputs", (DL_FUNC) &_hermiter_standardizeInputs, 4},
     {"_hermiter_standardizeInputsEW", (DL_FUNC) &_hermiter_standardizeInputsEW, 5},